Bluetooth: Classic: Move SCO APIs from internal to public header
Move SCO-related structures, enums, and function declarations from
the internal header `subsys/bluetooth/host/classic/sco_internal.h`
to the new public header `include/zephyr/bluetooth/classic/sco.h`.

This change exposes SCO APIs to applications and external modules,
enabling them to:
- Create and manage SCO connections
- Register SCO servers and connection callbacks
- Monitor HCI activity for SCO connections

The internal header now only contains internal helper functions for
SCO channel state management.
