+++ subsys/bluetooth/host/classic/br.c
@@ -10,6 +10,7 @@
 #include <zephyr/bluetooth/bluetooth.h>
 #include <zephyr/bluetooth/hci.h>
 #include <zephyr/bluetooth/buf.h>
+#include <zephyr/bluetooth/classic/sco.h>
 
 #include <common/bt_str.h>
