+++ subsys/bluetooth/host/classic/hfp_hf.c
@@ -19,6 +19,7 @@
 #include <zephyr/bluetooth/classic/rfcomm.h>
 #include <zephyr/bluetooth/classic/hfp_hf.h>
 #include <zephyr/bluetooth/classic/sdp.h>
+#include <zephyr/bluetooth/classic/sco.h>
 
 #include <host/hci_core.h>
 #include <host/conn_internal.h>
